# Dictionary access is saniztized and optimized.