--- a/Code/GraphMol/Bond.h
+++ b/Code/GraphMol/Bond.h
@@ -18,6 +18,7 @@
 #include <Query/QueryObjects.h>
 #include <RDGeneral/types.h>
 #include <GraphMol/details.h>
+#include <boost/foreach.hpp>
 
 namespace RDKit{
   class ROMol;
@@ -303,15 +304,15 @@ namespace RDKit{
     }
     //! \overload
     template <typename T>
-    void setProp(const std::string key,T val, bool computed=false ) const{
+    void setProp(const std::string &key,T val, bool computed=false ) const{
       //setProp(key.c_str(),val);
       if (computed) {
-	STR_VECT compLst;
-	if(hasProp(detail::computedPropName)) getProp(detail::computedPropName, compLst);
-	if (std::find(compLst.begin(), compLst.end(), key) == compLst.end()) {
-	  compLst.push_back(key);
-	  dp_props->setVal(detail::computedPropName, compLst);
-	}
+          STR_VECT compLst;
+          getPropIfPresent(detail::computedPropName, compLst);
+          if (std::find(compLst.begin(), compLst.end(), key) == compLst.end()) {
+              compLst.push_back(key);
+              dp_props->setVal(detail::computedPropName, compLst);
+          }
       }
       dp_props->setVal(key,val);
     }
@@ -337,20 +338,33 @@ namespace RDKit{
     }
     //! \overload
     template <typename T>
-    void getProp(const std::string key,T &res) const {
+    void getProp(const std::string &key,T &res) const {
       PRECONDITION(dp_props,"getProp called on empty property dict");
       dp_props->getVal(key,res);
-    
     }
-    //! \overload
+
+    //! \Overload
     template <typename T>
     T getProp(const char *key) const {
       return dp_props->getVal<T>(key);
     }
     //! \overload
     template <typename T>
-    T getProp(const std::string key) const {
+    T getProp(const std::string &key) const {
       return dp_props->getVal<T>(key);
+    }
+
+    //! returns whether or not we have a \c property with name \c key
+    //!  and assigns the value if we do
+
+    template <typename T>
+    bool getPropIfPresent(const char *key,T &res) const {
+        return dp_props->getValIfPresent(key,res);
+    }
+    //! \overload
+    template <typename T>
+    bool getPropIfPresent(const std::string &key,T &res) const {
+        return dp_props->getValIfPresent(key,res);
     }
 
     //! returns whether or not we have a \c property with name \c key
@@ -359,7 +373,7 @@ namespace RDKit{
       return dp_props->hasVal(key);
     };
     //! \overload
-    bool hasProp(const std::string key) const {
+    bool hasProp(const std::string &key) const {
       if(!dp_props) return false;
       return dp_props->hasVal(key);
     };
@@ -377,10 +391,9 @@ namespace RDKit{
       clearProp(what);
     };
     //! \overload
-    void clearProp(const std::string key) const {
-      if(hasProp(detail::computedPropName)){
-	STR_VECT compLst;
-	getProp(detail::computedPropName, compLst);
+    void clearProp(const std::string &key) const {
+      STR_VECT compLst;
+      if (getPropIfPresent(detail::computedPropName, compLst)) {
 	STR_VECT_I svi = std::find(compLst.begin(), compLst.end(), key);
 	if (svi != compLst.end()) {
 	  compLst.erase(svi);
@@ -388,19 +401,18 @@ namespace RDKit{
 	}
       }
       dp_props->clearVal(key);
-    };
+    }
 
     //! clears all of our \c computed \c properties
     void clearComputedProps() const {
-      if(!hasProp(detail::computedPropName)) return;
       STR_VECT compLst;
-      getProp(detail::computedPropName, compLst);
-      STR_VECT_CI svi;
-      for (svi = compLst.begin(); svi != compLst.end(); svi++) {
-	dp_props->clearVal(*svi);
+      if(getPropIfPresent(detail::computedPropName, compLst)) {
+	BOOST_FOREACH(const std::string &sv,compLst){
+	  dp_props->clearVal(sv);
+	}
+	compLst.clear();
+	dp_props->setVal(detail::computedPropName, compLst);
       }
-      compLst.clear();
-      dp_props->setVal(detail::computedPropName, compLst);
     }
 
     //! calculates any of our lazy \c properties