--- a/Code/GraphMol/Chirality.h
+++ b/Code/GraphMol/Chirality.h
@@ -26,7 +26,7 @@ namespace RDKit{
                     Should be at least mol.getNumAtoms() long.
     
       <b>Notes:</b>
-         - All atoms gain a property "_CIPRank" with their overall
+         - All atoms gain a property common_properties::_CIPRank with their overall
            CIP ranking.
     
     */